move checkForAtomicOperators to collection_ops
