+++ lib/collection.js
@@ -20,6 +20,7 @@ const getReadPreference = require('./utils').getReadPreference;
 
 // Operations
 const bulkWrite = require('./operations/collection_ops').bulkWrite;
+const checkForAtomicOperators = require('./operations/collection_ops').checkForAtomicOperators;
 const count = require('./operations/collection_ops').count;
 const createIndex = require('./operations/collection_ops').createIndex;
 const createIndexes = require('./operations/collection_ops').createIndexes;
@@ -722,19 +723,6 @@ Collection.prototype.updateOne = function(filter, update, options, callback) {
   }
 
   return executeOperation(this.s.topology, updateOne, [this, filter, update, options, callback]);
-};
-
-var checkForAtomicOperators = function(update) {
-  const keys = Object.keys(update);
-
-  // same errors as the server would give for update doc lacking atomic operators
-  if (keys.length === 0) {
-    return toError('The update operation document must contain at least one atomic operator.');
-  }
-
-  if (keys[0][0] !== '$') {
-    return toError('the update operation document must contain atomic operators.');
-  }
 };
 
 /**
