+++ lib/operations/collection_ops.js
@@ -147,6 +147,20 @@ function bulkWrite(coll, operations, options, callback) {
   });
 }
 
+// Check the update operation to ensure it has atomic operators.
+function checkForAtomicOperators(update) {
+  const keys = Object.keys(update);
+
+  // same errors as the server would give for update doc lacking atomic operators
+  if (keys.length === 0) {
+    return toError('The update operation document must contain at least one atomic operator.');
+  }
+
+  if (keys[0][0] !== '$') {
+    return toError('the update operation document must contain atomic operators.');
+  }
+}
+
 /**
  * Count the number of documents in the collection that match the query.
  *
@@ -1378,6 +1392,7 @@ function updateOne(coll, filter, update, options, callback) {
 
 module.exports = {
   bulkWrite,
+  checkForAtomicOperators,
   count,
   createIndex,
   createIndexes,
